requirements.txt
google-genai>=1.55.0
python-dotenv
django
gunicorn
streamlit
flask
